Getting demo going in webdrivermanager Made base config code reusable
