+++ cubano-webdriver-manager/config.properties
@@ -1,0 +1,82 @@
+# Usage notes
+# -----------
+# Properties can come in two flavors: 
+#  * environment specific - ie prefixed with name of environment, eg 'dev.prop.value = something' 
+#  * default value - ie anything without an envrionment prefix, eg 'prop.value = something'
+# 
+# These values can be also overridden on a per user basis by placing custom settings in a user.properties file.
+#
+# The search order for looking up a configuration property is:
+#  1. user.properties file
+#     1.1 requested property prefixed with '<environment>.'
+#     1.2 requested property 
+#  2. config.properties file
+#     2.1 requested property prefixed with '<environment>.'
+#     2.2 requested property 
+#
+#
+# User Properties
+# ---------------
+# Create file 'user.properties'
+# Enter required settings, the following may be a useful starting point:
+#
+# environment = ?
+#
+# webdriver.firefox.activatePlugins = true
+# webdriver.firefox.exe = C:/NotBackedUp/Apps/FirefoxPortable/FF.exe
+# 
+# General Settings
+# ----------------
+# 	environment 				Specifies which section of the configuration file to get the settings for.  
+#								NOTE: If a system property of 'environment' has been set then this will be used, otherwise it will use the default 
+#									  setting.  If the default setting is not supplied an exception will be thrown.
+#
+# WebDriver
+# ---------
+#	webdriver.browser			browser to test against:
+#									local: FireFox, Chrome, IE 
+#									remote: as above but add browser version, see the BrowserStack class for list of supported browsers/devices
+#	webdriver.defaultTimeout	default timeout value to use on element look ups
+#	webdriver.browserSize		Specify a custom window size for browser, if not specified is maximised 
+#	webdriver.<browser>.exe 	location of browser, eg: %USERPROFILE%/Documents/Mozilla FireFox Portable/FirefoxPortable.exe
+#	webdriver.<browser>.activatePlugins	
+#								If value is 'true' will add FireBug and FirePath plugins (only supported for FireFox currently)
+#									WARNING: do not activate this by default - it should not be enabled on CI server as it is only 
+#											 for use by the test developers
+#									To update plugins goto below locations and download to libs project folder: 
+#										* https://addons.mozilla.org/en-US/firefox/addon/firebug/
+#										* https://addons.mozilla.org/en-US/firefox/addon/firepath/ 
+#	webdriver.timeouts.implicitlywait	
+#								If choosing to use implicit waits using @timeout provided by Yandex HtmlElements can use this setting
+#								to set a global default wait period on all WebElements.  Defaults to zero.  
+#								WARNING: Do not mix with Selenium WebDriver's implicit or explicit waits as the timeout behaviour
+#										 becomes unpredictable.
+#
+# Environment
+# -----------
+#	<env>.<property>			Depends on the application you are testing
+#
+#
+# Proxy Settings
+# --------------
+# If accessing a system outside of a corporate proxy you may need to supply username and password information,
+# This information should not go into the config.properties file as this will get checked in.  Instead each user
+# should enter the following settings into the user.properties file and ensure that user.properties is exlcuded from
+# source control using svn or git ignore settings:
+# 
+#    proxy.domain = ?
+#    proxy.username = ?
+#    proxy.password = ?
+#
+# The following proxy settings can sit in config.properties or user.properties:
+#    proxy.required = true/false
+#    proxy.host = nwsa001
+#    proxy.port = 8080
+
+############### Default Values ###############
+
+environment = dev
+
+webdriver.browser = firefox
+webdriver.defaultTimeout = 5
+webdriver.firefox.activatePlugins = false
