+++ cubano-webdriver-manager/src/test/resources/demo/HelloWorld.md
@@ -1,0 +1,7 @@
+# Hello World
+
+## [Example Failure](-)
+[Hello World!](- "?=getGreetingFailure()")
+
+## [Example Exception](-)
+[Hello World!](- "?=getGreetingException()")
