+++ cubano-webdriver/src/main/java/org/concordion/cubano/driver/web/config/WebDriverConfig.java
@@ -1,13 +1,9 @@
 package org.concordion.cubano.driver.web.config;
 
-import java.util.Enumeration;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Properties;
 
-import org.concordion.cubano.utils.CaseSensitiveConfigLoader;
-import org.concordion.cubano.utils.ConfigLoader;
-import org.concordion.cubano.utils.DefaultConfigLoader;
+import org.concordion.cubano.utils.Config;
 
 /**
  * Reads and supplies properties from the <code>config.properties</code> file that are required by the framework.
@@ -21,12 +17,7 @@ import org.concordion.cubano.utils.DefaultConfigLoader;
  *
  * @author Andrew Sumner
  */
-public class WebDriverConfig {
-    private Properties properties;
-    private Properties userProperties = null;
-
-    // Environment
-    private String environment = null;
+public class WebDriverConfig extends Config {
 
     // Browser 
     private String browserProvider;
@@ -45,6 +36,18 @@ public class WebDriverConfig {
     private String proxyUsername;
     private String proxyPassword;
 
+    protected WebDriverConfig() {
+    	super();
+    }
+    
+    protected WebDriverConfig(Properties properties) {
+        super(properties);
+    }
+
+    protected WebDriverConfig(Properties properties, Properties userProperties) {
+        super(properties, userProperties);
+    }
+       
     private static class WDCHolder {
         static final WebDriverConfig INSTANCE = new WebDriverConfig();
     }
@@ -53,39 +56,7 @@ public class WebDriverConfig {
         return WDCHolder.INSTANCE;
     }
 
-    /**
-     * Prevent this class from being constructed.
-     */
-    protected WebDriverConfig() {
-        this(new DefaultConfigLoader());
-    }
-
-    /**
-     * Prevent this class from being constructed.
-     */
-    protected WebDriverConfig(ConfigLoader loader) {
-        this(loader.getProperties(), loader.getUserProperties());
-    }
-
-    protected WebDriverConfig(Properties properties) {
-        this(properties, null);
-    }
-
-    protected WebDriverConfig(Properties properties, Properties userProperties) {
-        this.properties = properties;
-        this.userProperties = userProperties;
-
-        loadCommonProperties();
-    }
-
-    private void loadCommonProperties() {
-        // Jenkins might supply value
-        environment = System.getProperty("environment", "").toLowerCase();
-
-        if (environment.isEmpty()) {
-            environment = getProperty("environment");
-        }
-
+    protected void loadProperties() {
         // Browser
         browserProvider = getOptionalProperty("webdriver.browserprovider", "org.concordion.cubano.driver.web.provider.LocalBrowserProvider");
         browserType = System.getProperty("browser");
@@ -123,128 +94,6 @@ public class WebDriverConfig {
 		}
     }
 
-	protected Map<String, String> getPropertiesStartingWith(String keyPrefix) {
-		Map<String, String> result = new HashMap<>();
-
-		searchPropertiesFrom(new CaseSensitiveConfigLoader().getProperties(), keyPrefix, result);
-		searchPropertiesFrom(new CaseSensitiveConfigLoader().getUserProperties(), keyPrefix, result);
-
-		return result;
-	}
-
-	private void searchPropertiesFrom(Properties properties, String keyPrefix, Map<String, String> result) {
-		@SuppressWarnings("unchecked")
-		Enumeration<String> en = (Enumeration<String>) properties.propertyNames();
-		while (en.hasMoreElements()) {
-			String propName = en.nextElement();
-			String propValue = properties.getProperty(propName);
-
-			if (propName.startsWith(keyPrefix)) {
-				result.put(propName, propValue);
-			}
-		}
-	}
-
-    /**
-     * Get the property for the current environment, if that is not found it will look for "default.{@literal <key>}".
-     *
-     * @param key Id of the property to look up
-     * @param isRequired true if the property is mandatory, throws RuntimeException if true and property not present
-     * @return Property value if found, throws exception if not found
-     */
-    protected String getProperty(String key, boolean isRequired) {
-        String value = retrieveProperty(key);
-
-        if (isRequired && value.isEmpty()) {
-            throw new RuntimeException(String.format("Unable to find property %s", key));
-        }
-
-        return value;
-    }
-
-    /**
-     * Get the property for the current environment, if that is not found it will look for "default.{@literal <key>}".
-     *
-     * @param key Id of the property to look up
-     * @return Property value if found, throws exception if not found
-     */
-    protected String getProperty(String key) {
-        return getProperty(key, false);
-    }
-
-    /**
-     * Get the property for the current environment, if that is not found it will look for "default.{@literal <key>}".
-     *
-     * @param key Id of the property to look up
-     * @return Property value if found, empty string if not found
-     */
-    protected String getOptionalProperty(String key) {
-        return retrieveProperty(key);
-    }
-
-    /**
-     * Get the property for the current environment, if that is not found it will look for "default.{@literal <key>}".
-     *
-     * @param key          Id of the property to look up
-     * @param defaultValue value to use if property is not found
-     * @return Property value if found, defaultValue if not found
-     */
-    protected String getOptionalProperty(String key, String defaultValue) {
-        String value = retrieveProperty(key);
-
-        if (value.isEmpty()) {
-            return defaultValue;
-        }
-
-        return value;
-    }
-
-    private String retrieveProperty(String key) {
-        String value = null;
-
-        // prefix = System.getProperty("user.name").toLowerCase();
-        if (userProperties != null) {
-            value = retrievePropertyFrom(userProperties, key);
-        }
-
-        if (value == null) {
-            value = retrievePropertyFrom(properties, key);
-        }
-
-        if (value == null) {
-            value = "";
-        }
-
-        return value;
-    }
-
-    private String retrievePropertyFrom(Properties properties, String key) {
-        String value = null;
-
-        // Attempt to get setting for environment
-        if (environment != null && !environment.isEmpty()) {
-            value = properties.getProperty(environment + "." + key);
-        }
-
-        // Attempt to get default setting
-        if (value == null) {
-            value = properties.getProperty(key);
-        }
-
-        if (value != null) {
-            value = value.trim();
-        }
-
-        return value;
-    }
-
-    /**
-     * @return Configured environment.
-     */
-    public String getEnvironment() {
-        return environment;
-    }
-
     // Browser
     private boolean useLocalBrowser() {
         return !browserType.contains(" ");
@@ -269,16 +118,6 @@ public class WebDriverConfig {
         }
 
         return "";
-    }
-
-    /**
-     * Activate developer plugins - FireFox only browser supported currently and will add FireBug and FirePath.
-     *
-     * @return true or false
-     * @deprecated use shouldActivatePlugins
-     */
-    public boolean activatePlugins() {
-        return shouldActivatePlugins();
     }
 
     /**
